requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.9
Pillow==9.5.0
django-admin-interface==0.24.0
django-colorfield==0.9.0
django-cors-headers==4.1.0
django-filter==23.2
python-dotenv==1.0.0
Markdown==3.4.3
drf-yasg==1.21.7
whitenoise==6.5.0
gunicorn==21.2.0
gevent==23.9.1
